Huge Bug-fixes + basis for User accounts
Somehow, in the game logic module, the tests were duplicated and not working properly. I had to delete them and fix other issues with the remaining tests. Also, I must have accidentally added an unfinished class to the commit and pushed it. As a result, after cloning the project, it wasn't working properly. I apologize if you cloned this buggy version of the project. For good news I've just started working on user accounts.